move useFetchData-related files into own directory
for consistency
